fix some warnings

+++ src/vm/optimize.c
@@ -439,7 +439,7 @@ UserFunction *access_vars_via_refslots(UserFunction *uf) {
             bool continue_outer = false;
             for (int k = 0; k < info[obj_slot].fields_len; ++k) {
               const char *name = info[obj_slot].names_ptr[k];
-              int name_len = info[obj_slot].names_len_ptr[k];
+              size_t name_len = (size_t) info[obj_slot].names_len_ptr[k];
               if (keylen == name_len && strncmp(keyptr, name, keylen) == 0) {
                 int refslot = ref_slots_ptr[obj_slot][k];
                 use_range_start(&builder, *instr_belongs_to_p(&uf->body, instr));
